move fn to utility
